lib/nucleus_one_dart_sdk.dart - make several types available, and make constructors available for testing, not exhaustive lib/src/api_model/document.dart lib/src/api_model/document.g.dart lib/src/model/document.dart - fix crash due to paper margins assumed to be int, but can be double lib/src/http.dart - _executeStandardHttpRequest method; ensuring HttpClient is properly closed lib/src/model/document_package.dart - approval field is no longer required lib/src/model/document_subscription_for_client.dart - fromApiModel method; added field defaults lib/src/model/organization_project.dart lib/src/model/user_organization.dart lib/src/model/user_organization_project.dart - Added @visibleForTesting constructor lib/src/nucleus_one.dart - remove @visibleForTesting from NucleusOneApp constructor
--------------------------------------------------

commit 968abd51d3a29ff555dbfbb779032c333b9fa3e3 (HEAD -> my-ApiMappingRefresh, origin/my-ApiMappingRefresh)
Author: jibbers42 <797824+jibbers42@users.noreply.github.com>
Date:   Wed Aug 31 20:57:57 2022 -0400

    make several types available, and make constructors available for testing, not ehaustive

commit 362230998d109713c695255706f7a276c5a48fe7
Author: jibbers42 <797824+jibbers42@users.noreply.github.com>
Date:   Wed Aug 31 20:56:10 2022 -0400

    fix crash due to page margins assumed to be int, but can be double

commit fc5b0ecd4c5ad2301ed21b0531ee6fddf35d512d
Author: jibbers42 <797824+jibbers42@users.noreply.github.com>
Date:   Wed Aug 31 20:54:48 2022 -0400

    remove @visibleForTesting from NucleusOneApp constructor

commit 861c5e8793750ab69190c61b14586b4e1d4e4f63
Author: jibbers42 <797824+jibbers42@users.noreply.github.com>
Date:   Wed Aug 31 20:54:05 2022 -0400

    make n1 HttpException visible to get http status code on error

commit d703be883ec44d69419404a1617f50eafded48bf
Author: jibbers42 <797824+jibbers42@users.noreply.github.com>
Date:   Wed Aug 31 20:49:41 2022 -0400

    workarounds for a few properties being null and crashing

commit 437bda05c36fc6dbde510b7f4dbd95fa9fbf06be
Author: jibbers42 <797824+jibbers42@users.noreply.github.com>
Date:   Wed Aug 31 20:44:27 2022 -0400

    close http client to keep cli from hanging at exit

    This may not be a great idea. I imagine it's possible dart keeps
    connections open for efficiency regarding multiple calls to the same
    server. It may be better to have a close() methos or similar on the sdk
    that can be called at app exit.

+++ lib/src/api_model/document.dart
@@ -127,16 +127,16 @@ class Document extends Entity {
   String? paperOrientation;
 
   @JsonKey(name: 'PaperMarginLeft')
-  int? paperMarginLeft;
+  double? paperMarginLeft;
 
   @JsonKey(name: 'PaperMarginRight')
-  int? paperMarginRight;
+  double? paperMarginRight;
 
   @JsonKey(name: 'PaperMarginTop')
-  int? paperMarginTop;
+  double? paperMarginTop;
 
   @JsonKey(name: 'PaperMarginBottom')
-  int? paperMarginBottom;
+  double? paperMarginBottom;
 
   // coverage:ignore-start
   /// [toJson] is the convention for a class to declare support for serialization
